+++ light-base/src/json_rpc_service/background/chain_head.rs
@@ -926,8 +926,8 @@ impl<TPlat: PlatformRef> ChainHeadFollowTask<TPlat> {
                                 methods::ChainHeadStorageType::Hash => {
                                     sync_service::StorageRequestItemTy::Hash
                                 }
-                                methods::ChainHeadStorageType::ClosestAncestorMerkleValue => {
-                                    sync_service::StorageRequestItemTy::ClosestAncestorMerkleValue
+                                methods::ChainHeadStorageType::ClosestDescendantMerkleValue => {
+                                    sync_service::StorageRequestItemTy::ClosestDescendantMerkleValue
                                 }
                                 methods::ChainHeadStorageType::DescendantsValues => {
                                     sync_service::StorageRequestItemTy::DescendantsValues
@@ -975,7 +975,6 @@ impl<TPlat: PlatformRef> ChainHeadFollowTask<TPlat> {
                                             value: Some(methods::HexString(value?)),
                                             hash: None,
                                             merkle_value: None,
-                                            merkle_value_key: None,
                                         })
                                     }
                                     sync_service::StorageResultItem::Hash { key, hash } => {
@@ -984,7 +983,6 @@ impl<TPlat: PlatformRef> ChainHeadFollowTask<TPlat> {
                                             value: None,
                                             hash: Some(methods::HexString(hash?.to_vec())),
                                             merkle_value: None,
-                                            merkle_value_key: None,
                                         })
                                     }
                                     sync_service::StorageResultItem::DescendantValue { key, value, .. } => {
@@ -993,7 +991,6 @@ impl<TPlat: PlatformRef> ChainHeadFollowTask<TPlat> {
                                             value: Some(methods::HexString(value)),
                                             hash: None,
                                             merkle_value: None,
-                                            merkle_value_key: None,
                                         })
                                     }
                                     sync_service::StorageResultItem::DescendantHash { key, hash, .. } => {
@@ -1002,17 +999,14 @@ impl<TPlat: PlatformRef> ChainHeadFollowTask<TPlat> {
                                             value: None,
                                             hash: Some(methods::HexString(hash.to_vec())),
                                             merkle_value: None,
-                                            merkle_value_key: None,
                                         })
                                     }
-                                    sync_service::StorageResultItem::ClosestAncestorMerkleValue { requested_key, merkle_value } => {
-                                        let (merkle_value_of, merkle_value) = merkle_value?;
+                                    sync_service::StorageResultItem::ClosestDescendantMerkleValue { requested_key, merkle_value } => {
                                         Some(methods::ChainHeadStorageResponseItem {
                                             key: methods::HexString(requested_key),
                                             value: None,
                                             hash: None,
-                                            merkle_value: Some(methods::HexString(merkle_value)),
-                                            merkle_value_key: Some(format!("0x{}", merkle_value_of.iter().map(|n| format!("{:x}", n)).collect::<String>())),
+                                            merkle_value: Some(methods::HexString(merkle_value?)),
                                         })
                                     }
                                 })