+++ light-base/src/sync_service.rs
@@ -413,7 +413,7 @@ impl<TPlat: PlatformRef> SyncService<TPlat> {
                 key: Vec<u8>,
                 hash: bool,
             },
-            ClosestAncestorMerkleValue {
+            ClosestDescendantMerkleValue {
                 key: Vec<u8>,
             },
         }
@@ -440,8 +440,8 @@ impl<TPlat: PlatformRef> SyncService<TPlat> {
                     key: request.key,
                     hash: true,
                 },
-                StorageRequestItemTy::ClosestAncestorMerkleValue => {
-                    RequestImpl::ClosestAncestorMerkleValue { key: request.key }
+                StorageRequestItemTy::ClosestDescendantMerkleValue => {
+                    RequestImpl::ClosestDescendantMerkleValue { key: request.key }
                 }
             })
             .collect::<Vec<_>>();
@@ -494,8 +494,13 @@ impl<TPlat: PlatformRef> SyncService<TPlat> {
                         RequestImpl::ValueOrHash { key, .. } => {
                             keys.insert(key.clone());
                         }
-                        RequestImpl::ClosestAncestorMerkleValue { key } => {
-                            keys.insert(key.clone());
+                        RequestImpl::ClosestDescendantMerkleValue { key } => {
+                            // We query the parent of `key`.
+                            if key.is_empty() {
+                                keys.insert(Vec::new());
+                            } else {
+                                keys.insert(key[..key.len() - 1].to_owned());
+                            }
                         }
                     }
                 }
@@ -651,26 +656,23 @@ impl<TPlat: PlatformRef> SyncService<TPlat> {
                             }
                         }
                     }
-                    RequestImpl::ClosestAncestorMerkleValue { key } => {
-                        match decoded_proof.closest_ancestor_merkle_value(
+                    RequestImpl::ClosestDescendantMerkleValue { key } => {
+                        match decoded_proof.closest_descendant_merkle_value(
                             main_trie_root_hash,
                             &trie::bytes_to_nibbles(key.iter().copied()).collect::<Vec<_>>(),
                         ) {
-                            Ok(Some((ancestor_key, merkle_value))) => {
-                                final_results.push(StorageResultItem::ClosestAncestorMerkleValue {
+                            Ok(Some(merkle_value)) => final_results.push(
+                                StorageResultItem::ClosestDescendantMerkleValue {
                                     requested_key: key,
-                                    merkle_value: Some((
-                                        ancestor_key.to_vec(),
-                                        merkle_value.as_ref().to_vec(),
-                                    )),
-                                })
-                            }
-                            Ok(None) => {
-                                final_results.push(StorageResultItem::ClosestAncestorMerkleValue {
+                                    merkle_value: Some(merkle_value.as_ref().to_vec()),
+                                },
+                            ),
+                            Ok(None) => final_results.push(
+                                StorageResultItem::ClosestDescendantMerkleValue {
                                     requested_key: key,
                                     merkle_value: None,
-                                })
-                            }
+                                },
+                            ),
                             Err(proof_decode::IncompleteProofError { .. }) => {
                                 outcome_errors.push(StorageQueryErrorDetail::MissingProofEntry);
                             }
@@ -775,10 +777,10 @@ pub enum StorageRequestItemTy {
 
     /// The Merkle value of the trie node that is the closest ancestor to
     /// [`StorageRequestItem::key`] is requested.
-    /// A [`StorageResultItem::ClosestAncestorMerkleValue`] will be returned where
-    /// [`StorageResultItem::ClosestAncestorMerkleValue::requested_key`] is equal to
+    /// A [`StorageResultItem::ClosestDescendantMerkleValue`] will be returned where
+    /// [`StorageResultItem::ClosestDescendantMerkleValue::requested_key`] is equal to
     /// [`StorageRequestItem::key`].
-    ClosestAncestorMerkleValue,
+    ClosestDescendantMerkleValue,
 }
 
 /// An item returned by [`SyncService::storage_query`].
@@ -818,11 +820,11 @@ pub enum StorageResultItem {
         /// Hash of the storage value associated with [`StorageResultItem::DescendantHash::key`].
         hash: [u8; 32],
     },
-    /// Corresponds to a [`StorageRequestItemTy::ClosestAncestorMerkleValue`].
-    ClosestAncestorMerkleValue {
+    /// Corresponds to a [`StorageRequestItemTy::ClosestDescendantMerkleValue`].
+    ClosestDescendantMerkleValue {
         /// Key that was requested. Equal to the value of [`StorageRequestItem::key`].
         requested_key: Vec<u8>,
-        merkle_value: Option<(Vec<trie::Nibble>, Vec<u8>)>,
+        merkle_value: Option<Vec<u8>>,
     },
 }